pyproject.toml
[tool.poetry]
name = "protocol-management-system"
version = "0.1.0"
description = ""
authors = ["Adrian Koczur"]
readme = "README.md"
packages = [{include = "protocol_management_system"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.1"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
django-celery-beat = "^2.4.0"
flower = "^1.2.0"
psycopg2 = "^2.9.6"
django-extensions = "^3.2.1"
djangorestframework = "^3.14.0"
Pygments = "^2.15.1"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
Pillow = "^9.5.0"
python-barcode = "^0.14.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
isort = "^5.10.1"
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
